Style cleanup and relocate NotFoundException to http-service package
